sort out the RSA-PSS params passing to NSS
It turns out the changes made in 0d8680d78131b19f739e98438d7a9b06bdaf3cf0
weren't sound because they were returning a pointer to a local stack-allocated
SECItem. This fixes that.
Additionally, this introduces a SECItemMut type to differentiate between
SECItems with constant data that we pass as inputs to NSS functions and SECItems
with undefined data buffers that are to be filled out by NSS functions (as
output). It turns out that the unsafe call to set_len was actually what we
wanted in the first place. The paradigm of using Vec::with_capacity to allocate
an undefined buffer, passing that buffer to a foreign library, and then calling
set_len is the best approach.